optionsCalculator
This serves as a visual aid for planning an option strategy. remember, it's best to sell out of the money options rather than buy calls or puts unless you want high leverage or a specific type of strategy, like a black bird or an iron condor type. refer to someone who can speak the options language to help you plan any trades